feat(text-input-react): expose base input component
affects: @fremtind/jkl-text-input-react

Expose BaseInputField component to avoid possible issues with some client projects
