+++ packages/text-input-react/src/index.ts
@@ -1,2 +1,3 @@
 export { TextArea } from "./TextArea";
 export { TextInput } from "./TextInput";
+export { BaseInputField } from "./BaseInputField";
